fix :: usecase별로 분리 BACKEND-25
