+++ src/main/java/windeath44/orchestration/adapter/in/messaging/MemorialDeleteSubscribe.java
@@ -1,0 +1,18 @@
+package windeath44.orchestration.adapter.in.messaging;
+
+import com.example.avro.MemorialAvroSchema;
+import lombok.RequiredArgsConstructor;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.stereotype.Component;
+import windeath44.orchestration.domain.port.in.MemorialDeleteUseCase;
+
+@Component
+@RequiredArgsConstructor
+public class MemorialDeleteSubscribe  {
+  private final MemorialDeleteUseCase memorialDeleteUseCase;
+
+  @KafkaListener(topics = "character-memorializing-fail-response", groupId = "memorial")
+  public void compensate(MemorialAvroSchema message) {
+    memorialDeleteUseCase.execute(message);
+  }
+}
