+++ src/main/java/windeath44/orchestration/domain/port/in/MemorialDeleteUseCase.java
@@ -1,8 +1,7 @@
 package windeath44.orchestration.domain.port.in;
 
-
 import com.example.avro.MemorialAvroSchema;
 
-public interface CharacterMemorializingUseCase {
+public interface MemorialDeleteUseCase {
   void execute(MemorialAvroSchema memorialAvroSchema);
 }
